fix: improve parsing openapi enum keys (#12)
When parsing openapi spec, if we encounter an enum field, we take this
value unchanged. While you expect the string data type to be used. As a
result, we got an error.

These changes correct the situation. If a default value is set, it is
used. Otherwise, the first value from the enumeration is taken.

Tested locally. Helm parsing errors are no longer observed.