feat: use static image digest (#26)
During openapi spec generation, additional functions are used in the
modules that take the name of the image used and generate a digest for
it. We have no way of predicting the names for all the images we use.
This causes errors during spec generation due to missing values for
variables.

It is proposed to switch to using static values for all images in
templates. And also replace the generation call with a strict value.

To do this, after generating a chart, you are asked to go through the
prepared list of templates. And go through them searching. If an
occurrence of the string ':= include "helm_lib_module_' is found,
replace it with ':= "string"'. If an occurrence of the string 'image: '
is found, then replace it with 'image:
registry.example.com/module@sha256:d47...'